requirements.txt
Flask==2.2.2
flask_sqlalchemy==3.0.2
Flask-Migrate==4.0.2
PyMySQL==1.0.2
python-dotenv==0.21.1